+++ fiftyone/server/view.py
@@ -5,7 +5,6 @@ FiftyOne Server view.
 | `voxel51.com <https://voxel51.com/>`_
 |
 """
-import asyncio
 import strawberry as gql
 from typing import List, Optional
 
@@ -69,7 +68,7 @@ async def load_view(
 
             return view
 
-    return await fosu.run_sync_task(run)
+    return await fou.run_sync_task(run)
 
 
 def get_view(
